requirements.txt
nextcord>=2.1
function-cooldowns

flask
nest_asyncio

google_auth_oauthlib
google-api-python-client

pytz
parsedatetime
numerize

Pillow
matplotlib

aiohttp
bs4

asyncpg
mysql-connector-python

git+https://github.com/john-science/mazelib/